# public/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "PreCost API",
    "description": "Altınyıldız ön maliyet parametre ve sarf değeri yönetimi. Sezon / Marka / Kategori bazlı maliyet parametrelerini ve kombinasyon bazlı sarf değerlerini depolar.",
    "version": "1.0.0",
    "contact": {
      "name": "Altınyıldız IT"
    }
  },
  "host": "brprecost-58854d65e626.herokuapp.com",
  "basePath": "/",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {

    "/api/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Sistem sağlık kontrolü",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Sistem çalışıyor",
            "schema": {
              "type": "object",
              "properties": {
                "status":    { "type": "string", "example": "ok" },
                "timestamp": { "type": "string", "format": "date-time" }
              }
            }
          }
        }
      }
    },

    "/api/lookup/all": {
      "get": {
        "operationId": "getLookupAll",
        "summary": "Cache'lenmiş selector ve cost-attribute değer listelerini döner",
        "description": "Heroku DB'sindeki tüm lookup verisini döner. `stale=true` ise widget arka planda PLM'den yeniler.",
        "tags": ["Lookup"],
        "responses": {
          "200": {
            "description": "Lookup verisi",
            "schema": {
              "type": "object",
              "properties": {
                "lastRefreshedAt": { "type": "string", "format": "date-time" },
                "stale":           { "type": "boolean" },
                "attributes": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "role":     { "type": "string", "enum": ["selector","cost-attr"] },
                      "source":   { "type": "string", "enum": ["standard","extended"] },
                      "sourceId": { "type": "string" },
                      "label":    { "type": "string" },
                      "values": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id":   { "type": "string", "description": "PLM GlValId (standart) veya ExtFldDropDownId (extended)" },
                            "name": { "type": "string" },
                            "code": { "type": "string" },
                            "seq":  { "type": "integer" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/api/lookup/refresh": {
      "post": {
        "operationId": "refreshLookup",
        "summary": "Lookup cache'ini güncel veriyle yeniden yükler",
        "description": "Widget tarafı PLM'den çektiği selector ve cost-attr değerlerini bu endpoint'e gönderir; tüm lookup tabloları truncate edilip yeniden yazılır.",
        "tags": ["Lookup"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "required": ["selectors","costAttrs"],
            "properties": {
              "selectors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "key":      { "type": "string", "example": "Sezon" },
                    "sourceId": { "type": "string", "example": "58" },
                    "values":   {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id":   { "type": "string" },
                          "name": { "type": "string" },
                          "code": { "type": "string" },
                          "seq":  { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              },
              "costAttrs": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "key":      { "type": "string", "example": "Astar Tipi" },
                    "source":   { "type": "string", "enum": ["standard","extended"] },
                    "sourceId": { "type": "string", "example": "73" },
                    "values":   {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id":   { "type": "string" },
                          "name": { "type": "string" },
                          "code": { "type": "string" },
                          "seq":  { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              },
              "references": {
                "type": "array",
                "description": "Sadece cache'lenir, kullanıcıya gösterilmez (örn. Kumaş Eni). Widget 2 için.",
                "items": {
                  "type": "object",
                  "properties": {
                    "key":      { "type": "string", "example": "Kumaş Eni" },
                    "source":   { "type": "string", "enum": ["standard","extended"] },
                    "sourceId": { "type": "string", "example": "197" },
                    "values":   {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id":   { "type": "string" },
                          "name": { "type": "string" },
                          "code": { "type": "string", "description": "Kumaş Eni için cm değeri burada gelir" },
                          "seq":  { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "Yenileme başarılı",
            "schema": {
              "type": "object",
              "properties": {
                "ok":              { "type": "boolean" },
                "lastRefreshedAt": { "type": "string", "format": "date-time" },
                "count":           { "type": "integer" }
              }
            }
          },
          "400": { "description": "Eksik veya hatalı payload" }
        }
      }
    },

    "/api/ana-kumash/attributes": {
      "get": {
        "operationId": "getAnaKumashAttributes",
        "summary": "Maliyet parametresi olarak kullanılabilecek attribute listesi",
        "description": "Her attribute adı ve olası değerlerini döner. Değerler PLM GenericLookUp'tan beslenir.",
        "tags": ["Ana Kumaş"],
        "responses": {
          "200": {
            "description": "Attribute adı → değer listesi map'i",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "array",
                "items": { "type": "string" }
              },
              "example": {
                "Astar Tipi": ["Tam Astar", "Yarım Astar", "Astarsız"],
                "Desen Tipi": ["Düz", "Desenli", "Ekoseli"]
              }
            }
          }
        }
      }
    },

    "/api/ana-kumash/list": {
      "get": {
        "operationId": "listAnaKumashConfigs",
        "summary": "Kayıtlı tüm sezon/marka/kategori kombinasyonlarını listeler",
        "tags": ["Ana Kumaş"],
        "responses": {
          "200": {
            "description": "Kombinasyon listesi",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/ComboKey" }
            }
          }
        }
      }
    },

    "/api/ana-kumash/lookup-bulk": {
      "post": {
        "operationId": "lookupAnaKumashBulk",
        "summary": "Çoklu marka/kategori için config + values'ü tek istekte döndürür",
        "description": "Widget 2 (Cost Missing Models) için: tek bir sezon altında birden çok marka × kategori kombinasyonu için tanımlı maliyet parametreleri ve sarf değerlerini döner.",
        "tags": ["Ana Kumaş"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "required": ["season","brands","categories"],
            "properties": {
              "season":     { "type": "string", "example": "SS27" },
              "brands":     { "type": "array", "items": { "type": "string" }, "example": ["Beymen Business","Network"] },
              "categories": { "type": "array", "items": { "type": "string" }, "example": ["Ceket","Pantolon"] }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "Kombinasyon başına config ve values listesi",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "season":         { "type": "string" },
                  "brand":          { "type": "string" },
                  "styleCategory":  { "type": "string" },
                  "selectedAttrs":  { "type": "array", "items": { "type": "string" } },
                  "values": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "attrCombo":   { "type": "object", "additionalProperties": { "type": "string" } },
                        "consumption": { "type": "number" },
                        "unit":        { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Eksik veya hatalı payload" }
        }
      }
    },

    "/api/ana-kumash/config/{season}/{brand}/{styleCategory}": {
      "get": {
        "operationId": "getAnaKumashConfig",
        "summary": "Bir kombinasyon için seçili maliyet parametrelerini döner",
        "tags": ["Ana Kumaş"],
        "parameters": [
          { "in": "path", "name": "season",        "required": true, "type": "string", "example": "SS27" },
          { "in": "path", "name": "brand",          "required": true, "type": "string", "example": "Beymen Business" },
          { "in": "path", "name": "styleCategory",  "required": true, "type": "string", "example": "Ceket" }
        ],
        "responses": {
          "200": {
            "description": "Seçili attribute listesi",
            "schema": {
              "type": "object",
              "properties": {
                "selectedAttrs": {
                  "type": "array",
                  "items": { "type": "string" },
                  "example": ["Astar Tipi", "Desen Tipi"]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "saveAnaKumashConfig",
        "summary": "Bir kombinasyon için maliyet parametrelerini kaydeder",
        "description": "Attribute listesi değişmişse o kombinasyonun tüm sarf değerleri silinir.",
        "tags": ["Ana Kumaş"],
        "parameters": [
          { "in": "path", "name": "season",        "required": true, "type": "string" },
          { "in": "path", "name": "brand",          "required": true, "type": "string" },
          { "in": "path", "name": "styleCategory",  "required": true, "type": "string" },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["selectedAttrs"],
              "properties": {
                "selectedAttrs": {
                  "type": "array",
                  "items": { "type": "string" },
                  "example": ["Astar Tipi", "Desen Tipi"]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kayıt başarılı",
            "schema": {
              "type": "object",
              "properties": {
                "success":     { "type": "boolean" },
                "selectedAttrs": { "type": "array", "items": { "type": "string" } },
                "valuesReset": {
                  "type": "boolean",
                  "description": "true ise attribute listesi değişti ve eski sarf değerleri silindi"
                }
              }
            }
          }
        }
      }
    },

    "/api/ana-kumash/values/{season}/{brand}/{styleCategory}": {
      "get": {
        "operationId": "getAnaKumashValues",
        "summary": "Bir kombinasyon için kayıtlı sarf değerlerini döner",
        "tags": ["Ana Kumaş"],
        "parameters": [
          { "in": "path", "name": "season",        "required": true, "type": "string", "example": "SS27" },
          { "in": "path", "name": "brand",          "required": true, "type": "string", "example": "Beymen Business" },
          { "in": "path", "name": "styleCategory",  "required": true, "type": "string", "example": "Ceket" }
        ],
        "responses": {
          "200": {
            "description": "Sarf değerleri listesi",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/ConsumptionValue" }
            }
          }
        }
      },
      "post": {
        "operationId": "saveAnaKumashValues",
        "summary": "Bir kombinasyon için sarf değerlerini toplu kaydeder (upsert)",
        "description": "Gönderilen satırlar upsert edilir. Gönderilmeyen eski satırlar korunur.",
        "tags": ["Ana Kumaş"],
        "parameters": [
          { "in": "path", "name": "season",        "required": true, "type": "string" },
          { "in": "path", "name": "brand",          "required": true, "type": "string" },
          { "in": "path", "name": "styleCategory",  "required": true, "type": "string" },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["values"],
              "properties": {
                "values": {
                  "type": "array",
                  "items": { "$ref": "#/definitions/ConsumptionValueInput" }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kayıt başarılı",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "saved":   { "type": "integer", "description": "Kaydedilen satır sayısı" }
              }
            }
          }
        }
      }
    },

    "/api/cost-params/{costCategory}/list": {
      "get": {
        "operationId": "listCostParamsConfigs",
        "summary": "Bir maliyet kategorisi için kayıtlı tüm sezon/marka/kategori kombinasyonlarını listeler",
        "tags": ["Cost Params (generic)"],
        "parameters": [{ "in": "path", "name": "costCategory", "required": true, "type": "string", "enum": ["ana-kumash","astar-garni","iscilik","uretim-paket","malzeme"] }],
        "responses": { "200": { "description": "Kombinasyon listesi", "schema": { "type": "array", "items": { "$ref": "#/definitions/ComboKey" } } } }
      }
    },

    "/api/cost-params/{costCategory}/lookup-bulk": {
      "post": {
        "operationId": "lookupCostParamsBulk",
        "summary": "Bir maliyet kategorisi için çoklu marka × kategori bulk lookup",
        "tags": ["Cost Params (generic)"],
        "parameters": [
          { "in": "path", "name": "costCategory", "required": true, "type": "string" },
          { "in": "body", "name": "body", "required": true, "schema": {
            "type": "object", "required": ["season","brands","categories"],
            "properties": {
              "season":     { "type": "string" },
              "brands":     { "type": "array", "items": { "type": "string" } },
              "categories": { "type": "array", "items": { "type": "string" } }
            }
          } }
        ],
        "responses": { "200": { "description": "Combo başına config + values" } }
      }
    },

    "/api/cost-params/{costCategory}/config/{season}/{brand}/{styleCategory}": {
      "get": {
        "operationId": "getCostParamsConfig",
        "summary": "Bir kombinasyon için seçili maliyet parametrelerini döner",
        "tags": ["Cost Params (generic)"],
        "parameters": [
          { "in": "path", "name": "costCategory",   "required": true, "type": "string" },
          { "in": "path", "name": "season",          "required": true, "type": "string" },
          { "in": "path", "name": "brand",           "required": true, "type": "string" },
          { "in": "path", "name": "styleCategory",   "required": true, "type": "string" }
        ],
        "responses": { "200": { "description": "Seçili attribute listesi", "schema": { "type": "object", "properties": { "selectedAttrs": { "type": "array", "items": { "type": "string" } } } } } }
      },
      "post": {
        "operationId": "saveCostParamsConfig",
        "summary": "Bir kombinasyon için maliyet parametrelerini kaydeder",
        "description": "selectedAttrs boş array olabilir (örn. işçilikte parametre olmadan tek değer girilir). Liste değiştiyse o kombinasyonun değerleri silinir.",
        "tags": ["Cost Params (generic)"],
        "parameters": [
          { "in": "path", "name": "costCategory",   "required": true, "type": "string" },
          { "in": "path", "name": "season",          "required": true, "type": "string" },
          { "in": "path", "name": "brand",           "required": true, "type": "string" },
          { "in": "path", "name": "styleCategory",   "required": true, "type": "string" },
          { "in": "body", "name": "body", "required": true, "schema": {
            "type": "object", "required": ["selectedAttrs"],
            "properties": { "selectedAttrs": { "type": "array", "items": { "type": "string" } } }
          } }
        ],
        "responses": { "200": { "description": "Kayıt başarılı" } }
      }
    },

    "/api/cost-params/{costCategory}/values/{season}/{brand}/{styleCategory}": {
      "get": {
        "operationId": "getCostParamsValues",
        "summary": "Bir kombinasyon için kayıtlı değerleri döner",
        "tags": ["Cost Params (generic)"],
        "parameters": [
          { "in": "path", "name": "costCategory",   "required": true, "type": "string" },
          { "in": "path", "name": "season",          "required": true, "type": "string" },
          { "in": "path", "name": "brand",           "required": true, "type": "string" },
          { "in": "path", "name": "styleCategory",   "required": true, "type": "string" }
        ],
        "responses": { "200": { "description": "Değer listesi", "schema": { "type": "array", "items": { "$ref": "#/definitions/ConsumptionValue" } } } }
      },
      "post": {
        "operationId": "saveCostParamsValues",
        "summary": "Bir kombinasyon için değerleri toplu upsert eder",
        "tags": ["Cost Params (generic)"],
        "parameters": [
          { "in": "path", "name": "costCategory",   "required": true, "type": "string" },
          { "in": "path", "name": "season",          "required": true, "type": "string" },
          { "in": "path", "name": "brand",           "required": true, "type": "string" },
          { "in": "path", "name": "styleCategory",   "required": true, "type": "string" },
          { "in": "body", "name": "body", "required": true, "schema": {
            "type": "object", "required": ["values"],
            "properties": { "values": { "type": "array", "items": { "$ref": "#/definitions/ConsumptionValueInput" } } }
          } }
        ],
        "responses": { "200": { "description": "Kayıt başarılı" } }
      }
    }

  },

  "definitions": {

    "ComboKey": {
      "type": "object",
      "properties": {
        "season":        { "type": "string", "example": "SS27" },
        "brand":         { "type": "string", "example": "Beymen Business" },
        "styleCategory": { "type": "string", "example": "Ceket" }
      }
    },

    "ConsumptionValue": {
      "type": "object",
      "properties": {
        "attrCombo": {
          "type": "object",
          "description": "Attribute adı → değer map'i",
          "additionalProperties": { "type": "string" },
          "example": { "Astar Tipi": "Tam Astar", "Desen Tipi": "Düz" }
        },
        "consumption": {
          "type": "number",
          "format": "float",
          "description": "Kumaş sarfı (metre, 150cm en baz alınır)",
          "example": 2.1
        },
        "unit": { "type": "string", "example": "m" }
      }
    },

    "ConsumptionValueInput": {
      "type": "object",
      "required": ["attrCombo", "consumption"],
      "properties": {
        "attrCombo": {
          "type": "object",
          "additionalProperties": { "type": "string" },
          "example": { "Astar Tipi": "Tam Astar", "Desen Tipi": "Düz" }
        },
        "consumption": {
          "type": "number",
          "format": "float",
          "example": 2.1
        },
        "unit": { "type": "string", "default": "m" }
      }
    }

  }
}